feat: unify route condition resolver in expression language (#1821)
| Q             | A
| ------------- | ---
| Bug fix?      | no
| New feature?  | yes
| Doc PR?       | no
| License       | MIT

Unify route condition resolver in expression language. Since #1820 we can use `context.resolve()` in static route conditions, but this is not available if you use dynamic routing. With this PR you can use the same api with `context.getTenant()`. The old `resolver.resolve()` is not possible any more (BC-Break). All conditions moved to the tenancy bundle.